# frontend/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "WhaleHire API",
        "title": "WhaleHire API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/resume/list": {
            "get": {
                "description": "获取简历列表，支持分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resume"
                ],
                "summary": "获取简历列表",
                "operationId": "list-resume",
                "parameters": [
                    {
                        "type": "string",
                        "description": "下一页标识",
                        "name": "next_token",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "分页",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "每页多少条记录",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.ListResumeResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/resume/search": {
            "get": {
                "description": "根据关键词搜索简历",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resume"
                ],
                "summary": "搜索简历",
                "operationId": "search-resume",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "keywords",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "下一页标识",
                        "name": "next_token",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "分页",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "每页多少条记录",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.SearchResumeResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/resume/upload": {
            "post": {
                "description": "上传简历文件并进行解析",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resume"
                ],
                "summary": "上传简历",
                "operationId": "upload-resume",
                "parameters": [
                    {
                        "type": "file",
                        "description": "简历文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Resume"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/resume/{id}": {
            "get": {
                "description": "根据ID获取简历详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resume"
                ],
                "summary": "获取简历详情",
                "operationId": "get-resume",
                "parameters": [
                    {
                        "type": "string",
                        "description": "简历ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.ResumeDetail"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "更新简历信息，支持更新基本信息、教育经历、工作经历和技能",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resume"
                ],
                "summary": "更新简历",
                "operationId": "update-resume",
                "parameters": [
                    {
                        "type": "string",
                        "description": "简历ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新参数，支持更新基本信息和子表数据（教育经历、工作经历、技能）",
                        "name": "param",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateResumeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.Resume"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "删除简历",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resume"
                ],
                "summary": "删除简历",
                "operationId": "delete-resume",
                "parameters": [
                    {
                        "type": "string",
                        "description": "简历ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.Resp"
                        }
                    }
                }
            }
        },
        "/api/v1/resume/{id}/progress": {
            "get": {
                "description": "获取简历解析进度信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resume"
                ],
                "summary": "获取简历解析进度",
                "operationId": "get-resume-progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "简历ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.ResumeParseProgress"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/resume/{id}/reparse": {
            "post": {
                "description": "重新解析简历内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resume"
                ],
                "summary": "重新解析简历",
                "operationId": "reparse-resume",
                "parameters": [
                    {
                        "type": "string",
                        "description": "简历ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.Resp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.ListResumeResp": {
            "type": "object",
            "properties": {
                "has_next_page": {
                    "type": "boolean"
                },
                "next_token": {
                    "type": "string"
                },
                "resumes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Resume"
                    }
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
        "domain.Resume": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "current_city": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "highest_education": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parsed_at": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "resume_file_url": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.ResumeStatus"
                },
                "updated_at": {
                    "type": "integer"
                },
                "uploader_id": {
                    "type": "string"
                },
                "uploader_name": {
                    "type": "string"
                },
                "years_experience": {
                    "type": "number"
                }
            }
        },
        "domain.ResumeDetail": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "current_city": {
                    "type": "string"
                },
                "educations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ResumeEducation"
                    }
                },
                "email": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "experiences": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ResumeExperience"
                    }
                },
                "gender": {
                    "type": "string"
                },
                "highest_education": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ResumeLog"
                    }
                },
                "name": {
                    "type": "string"
                },
                "parsed_at": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "resume_file_url": {
                    "type": "string"
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ResumeSkill"
                    }
                },
                "status": {
                    "$ref": "#/definitions/domain.ResumeStatus"
                },
                "updated_at": {
                    "type": "integer"
                },
                "uploader_id": {
                    "type": "string"
                },
                "uploader_name": {
                    "type": "string"
                },
                "years_experience": {
                    "type": "number"
                }
            }
        },
        "domain.ResumeEducation": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "degree": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "major": {
                    "type": "string"
                },
                "resume_id": {
                    "type": "string"
                },
                "school": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
        "domain.ResumeExperience": {
            "type": "object",
            "properties": {
                "company": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "resume_id": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
        "domain.ResumeLog": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "created_at": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "resume_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
        "domain.ResumeParseProgress": {
            "type": "object",
            "properties": {
                "completed_at": {
                    "description": "完成解析时间",
                    "type": "string"
                },
                "error_message": {
                    "description": "错误信息（如果有）",
                    "type": "string"
                },
                "message": {
                    "description": "当前状态描述",
                    "type": "string"
                },
                "progress": {
                    "description": "解析进度百分比 0-100",
                    "type": "integer"
                },
                "resume_id": {
                    "type": "string"
                },
                "started_at": {
                    "description": "开始解析时间",
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.ResumeStatus"
                }
            }
        },
        "domain.ResumeSkill": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "level": {
                    "type": "string"
                },
                "resume_id": {
                    "type": "string"
                },
                "skill_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "integer"
                }
            }
        },
        "domain.ResumeStatus": {
            "type": "string",
            "enum": [
                "pending",
                "processing",
                "completed",
                "failed",
                "archived"
            ],
            "x-enum-comments": {
                "ResumeStatusArchived": "已归档",
                "ResumeStatusCompleted": "解析完成",
                "ResumeStatusFailed": "解析失败",
                "ResumeStatusPending": "已上传，等待解析",
                "ResumeStatusProcessing": "正在解析中"
            },
            "x-enum-varnames": [
                "ResumeStatusPending",
                "ResumeStatusProcessing",
                "ResumeStatusCompleted",
                "ResumeStatusFailed",
                "ResumeStatusArchived"
            ]
        },
...
        "domain.SearchResumeResp": {
            "type": "object",
            "properties": {
                "has_next_page": {
                    "type": "boolean"
                },
                "next_token": {
                    "type": "string"
                },
                "resumes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Resume"
                    }
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
        "domain.UpdateResumeEducation": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "操作类型：create, update, delete",
                    "type": "string"
                },
                "degree": {
                    "description": "学位",
                    "type": "string"
                },
                "end_date": {
                    "description": "结束时间",
                    "type": "string"
                },
                "id": {
                    "description": "教育经历ID，更新时必填",
                    "type": "string"
                },
                "major": {
                    "description": "专业",
                    "type": "string"
                },
                "school": {
                    "description": "学校",
                    "type": "string"
                },
                "start_date": {
                    "description": "开始时间",
                    "type": "string"
                }
            }
        },
        "domain.UpdateResumeExperience": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "操作类型：create, update, delete",
                    "type": "string"
                },
                "company": {
                    "description": "公司",
                    "type": "string"
                },
                "description": {
                    "description": "工作描述",
                    "type": "string"
                },
                "end_date": {
                    "description": "结束时间",
                    "type": "string"
                },
                "id": {
                    "description": "工作经历ID，更新时必填",
                    "type": "string"
                },
                "position": {
                    "description": "职位",
                    "type": "string"
                },
                "start_date": {
                    "description": "开始时间",
                    "type": "string"
                },
                "title": {
                    "description": "担任岗位",
                    "type": "string"
                }
            }
        },
        "domain.UpdateResumeReq": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "current_city": {
                    "type": "string"
                },
                "educations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.UpdateResumeEducation"
                    }
                },
                "email": {
                    "type": "string"
                },
                "experiences": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.UpdateResumeExperience"
                    }
                },
                "gender": {
                    "type": "string"
                },
                "highest_education": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.UpdateResumeSkill"
                    }
                },
                "years_experience": {
                    "type": "number"
                }
            }
        },
        "domain.UpdateResumeSkill": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "操作类型：create, update, delete",
                    "type": "string"
                },
                "description": {
                    "description": "技能描述",
                    "type": "string"
                },
                "id": {
                    "description": "技能ID，更新时必填",
                    "type": "string"
                },
                "level": {
                    "description": "技能水平",
                    "type": "string"
                },
                "skill_name": {
                    "description": "技能名称",
                    "type": "string"
                }
            }
        },
...
        "web.Resp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}